requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose==3.3.0
httpx==0.25.1
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
pymongo==4.6.1
motor==3.3.1
google-auth==2.27.0
requests==2.31.0
email-validator==2.1.0.post1
certifi==2024.2.2
mongoengine==0.27.0